# docs/fieldkit-server-api.openapi.yaml
openapi: 3.1.0
info:
  title: Fieldkit Device API
  description: |
    Versioned device API served from `https://fieldkit.infragilis.org`.
    This is the public contract the Fieldkit appliance sync client implements;
    it contains no server internals or secrets. Device access is authenticated
    with a device token linked to a server user. The manifest includes that user's
    personal files and subscribed shared data folders. The appliance only
    removes managed data files when its opt-in prune setting is enabled.
  version: 1.0.0
servers:
  - url: https://fieldkit.infragilis.org/api/v1
    description: Production (configurable at runtime)
security:
  - bearerAuth: []
paths:
  /device/update-latest:
    get:
      operationId: getDeviceUpdateLatest
      summary: Get the latest published appliance update
      responses:
        "200":
          description: Latest update metadata, or no available update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateLatest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Missing or invalid device token, or the token's user no longer exists.
  schemas:
...
    UpdateLatest:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
        version:
          type: string
        url:
          type: string
          format: uri
        sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        published_at:
          type: string
          format: date-time